+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,17 +1,12 @@
 package org.apache.coyote.http11;
 
-import com.techcourse.exception.BadRequestException;
 import com.techcourse.exception.UncheckedServletException;
 import com.techcourse.service.UserService;
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.net.Socket;
 import java.net.URL;
 import java.nio.file.Files;
-import java.util.HashMap;
 import java.util.Map;
 import org.apache.coyote.Processor;
 import org.apache.coyote.dto.ResourceResult;
@@ -22,15 +17,13 @@ public class Http11Processor implements Runnable, Processor {
 
     private static final Logger log = LoggerFactory.getLogger(Http11Processor.class);
     private static final String CRLF = "\r\n";
-    private static final String QUESTION = "?";
-    private static final String AND = "&";
-    private static final String EQUAL = "=";
-    private static final String COLON = ":";
 
     private final Socket connection;
+    private final HttpRequestReader httpRequestReader;
 
     public Http11Processor(final Socket connection) {
         this.connection = connection;
+        this.httpRequestReader = new HttpRequestReader();
     }
 
     @Override
@@ -44,17 +37,11 @@ public class Http11Processor implements Runnable, Processor {
         try (final var inputStream = connection.getInputStream();
              final var outputStream = connection.getOutputStream()) {
 
-            HttpRequest requestHeaders = readHttpRequest(inputStream);
-            RequestLine requestLine = requestHeaders.requestLine();
-            String method = requestLine.method();
-            String uri = requestLine.uri();
+            HttpRequest httpRequest = httpRequestReader.read(inputStream);
 
-            int questionIndex = uri.indexOf(QUESTION);
-            String path = findPath(questionIndex, uri);
-
-            Map<String, String> queries = findQueries(questionIndex, uri);
-
-            if (uri.startsWith("/login")) {
+            String path = httpRequest.path();
+            if (path.startsWith("/login")) {
+                Map<String, String> queries = httpRequest.queries();
                 String account = queries.get("account");
                 String password = queries.get("password");
                 UserService.login(account, password);
@@ -83,49 +70,6 @@ public class Http11Processor implements Runnable, Processor {
         }
     }
 
-    private HttpRequest readHttpRequest(final InputStream inputStream) throws IOException {
-        BufferedReader br = new BufferedReader(new InputStreamReader(inputStream));
-
-        RequestLine requestLine = readRequestLine(br);
-        Map<String, String> headers = readHeaders(br);
-
-        byte[] body = new byte[0];
-        String len = headers.get("content-length");
-        if (len != null) {
-            int contentLength = Integer.parseInt(len);
-            body = inputStream.readNBytes(contentLength);
-        }
-        return new HttpRequest(requestLine, headers, body);
-    }
-
-    private RequestLine readRequestLine(final BufferedReader br) throws IOException {
-        String requestLine = br.readLine();
-        if (requestLine == null || requestLine.isEmpty()) {
-            throw new BadRequestException("Request line is empty");
-        }
-
-        String[] parts = requestLine.split(" ");
-        String method = parts[0];
-        String uri = parts[1];
-        String version = parts[2];
-        return new RequestLine(method, uri, version);
-    }
-
-    private Map<String, String> readHeaders(final BufferedReader br) throws IOException {
-        Map<String, String> headers = new HashMap<>();
-        String line;
-        while ((line = br.readLine()) != null && !line.isEmpty()) {
-            int colon = line.indexOf(COLON);
-            if (colon < 0) {
-                throw new BadRequestException("Header is malformed");
-            }
-            String name = line.substring(0, colon).toLowerCase();
-            String value = line.substring(colon + 1);
-            headers.put(name, value);
-        }
-        return headers;
-    }
-
     private ResourceResult findResponseBody(String uri) throws IOException {
         if (uri.equals("/")) {
             return ResourceResult.found("Hello world!");
@@ -143,9 +87,6 @@ public class Http11Processor implements Runnable, Processor {
 
     private String findContentType(final String uri) {
         String encoding = ";charset=utf-8";
-        if (uri.endsWith(".html")) {
-            return "text/html" + encoding;
-        }
         if (uri.endsWith(".css")) {
             return "text/css" + encoding;
         }
@@ -155,37 +96,6 @@ public class Http11Processor implements Runnable, Processor {
         if (uri.endsWith(".png")) {
             return "image/png";
         }
-        return "";
-    }
-
-    private Map<String, String> parseQueryString(final int queryStartIndex, final String uri) {
-        if (!uri.contains(QUESTION)) {
-            return Map.of();
-        }
-        String query = uri.substring(queryStartIndex + 1);
-        String[] pairs = query.split(AND);
-        Map<String, String> queries = new HashMap<>();
-        for (String pair : pairs) {
-            int equalIndex = pair.indexOf(EQUAL);
-            String name = pair.substring(0, equalIndex);
-            String value = pair.substring(equalIndex + 1);
-            queries.put(name, value);
-        }
-        return queries;
-    }
-
-    private String findPath(final int queryStartIndex, final String uri) {
-        if (queryStartIndex < 0) {
-            return uri;
-        }
-        return uri.substring(0, queryStartIndex);
-    }
-
-    private Map<String, String> findQueries(final int questionIndex, final String uri) {
-        if (questionIndex > 0) {
-            String rawQueries = uri.substring(questionIndex + 1);
-            return parseQueryString(questionIndex, rawQueries);
-        }
-        return Map.of();
+        return "text/html" + encoding;
     }
 }
