+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequest.java
@@ -5,6 +5,8 @@ import java.util.Map;
 public record HttpRequest(
         RequestLine requestLine,
         Map<String, String> headers,
+        String path,
+        Map<String, String> queries,
         byte[] body
 ) {
     public String bodyAsString() {
